Added new "Show nested submodules" setting.
By default, the browse window shows nested submodules in the submodule
drop-down.  This is convenient for small repositories, but cumbersome
to use and slow for larger repositories:

* If the top project has a submodule with many submodules of its own,
  the background CPU usage will be high due to checking submodule status.

* If too many submodules are shown in the menu, it presents a usability
  problem as the list will scroll off the screen.

We add a setting to allow the user to disable checking of nested
submodules.  This will be useful to do for larger repositories with many
submodules.